Remove duplicate price storage

+++ src/libraries/MathLib.sol
@@ -133,7 +133,7 @@ library MathLib {
     }
 
     /// @notice Returns the largest of two numbers.
-    function max(uint256 a, uint256 b) internal pure returns (uint256) {
+    function max(uint96 a, uint96 b) internal pure returns (uint96) {
         return a < b ? b : a;
     }
 }
